graph-search/requirements.txt
google-api-python-client
google-cloud-enterpriseknowledgegraph